Rename new button to ButtonV2
