+++ packages/button/src/IconButton.tsx
@@ -9,7 +9,7 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import { spacingUnit } from '@ndla/core';
-import Button, { ButtonSize, ButtonProps } from './Button';
+import Button, { ButtonSize, ButtonProps } from './ButtonV2';
 
 interface StyledButtonProps extends ButtonProps {
   svgSize: number;
